requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
Markdown==3.3.4
Pillow==8.3.1
playwright==1.14.0
psycopg2==2.9.1
pyee==8.2.2
PyJWT==2.1.0
python-slugify==5.0.2
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
websockets==9.1
whitenoise==5.3.0
